Refactor extract objects to model directory
