+++ ctui/model/attribute.py
@@ -1,0 +1,8 @@
+class Attribute:
+
+    def __init__(self, key, value):
+        self.key = key
+        self.value = value
+
+    def __eq__(self, other):
+        return self.key == other.key and self.value == other.value
